🚚 refactor: Improve states folder structure
